+++ src/index.js
@@ -1,5 +1,8 @@
 export {default as DB, SQLite} from './DB'
 export {default as EventQueue} from './EventQueue'
-export {default as EventSourcingDB} from './EventSourcingDB'
+export {
+	default as EventSourcingDB,
+	applyResult,
+	ESModel,
+} from './EventSourcingDB'
 export {default as JsonModel} from './JsonModel'
-export {default as ESModel} from './EventSourcingDB/ESModel'